tests/test_parser_etl.py
import pytest
from pathlib import Path
from datetime import datetime
from etl_utils.parser import (
    parse_overview_data,
    parse_resort_page,
    parse_snow_forecast_images,
)

@pytest.fixture
def lelex_crozet_html():
    fixture_path = Path(__file__).parent / "fixtures" / "lelex-crozet.html"
    with open(fixture_path, "r", encoding="utf-8") as f:
        return f.read()

def test_parse_snow_forecast_images():
    """Test parsing of snow forecast images."""
    html = """
    <div class="snowforecast-img">
        <a href="https://vcdn.bergfex.at/images/resized/8b/daily.jpg" data-caption="Daily Caption">
            <img src="..." alt="Daily Alt">
        </a>
    </div>
    <div class="snowforecast-img">
        <a href="https://vcdn.bergfex.at/images/resized/5d/12h.jpg" data-caption="12h Caption">
            <img src="..." alt="12h Alt">
        </a>
    </div>
    <div class="snowforecast-img">
        <a href="https://vcdn.bergfex.at/images/resized/7b/summary.jpg" data-caption="Summary Caption">
            <img src="..." alt="Summary Alt">
        </a>
    </div>
    """
    
    # Test page 0
    result_page_0 = parse_snow_forecast_images(html, 0)
    assert result_page_0["daily_forecast_url"] == "https://vcdn.bergfex.at/images/resized/8b/daily.jpg"
    assert result_page_0["daily_caption"] == "Daily Caption"
    assert "summary_url" not in result_page_0
    
    # Test page 1
    result_page_1 = parse_snow_forecast_images(html, 1)
    assert result_page_1["daily_forecast_url"] == "https://vcdn.bergfex.at/images/resized/8b/daily.jpg"
    assert result_page_1["daily_caption"] == "Daily Caption"
    assert result_page_1["summary_url"] == "https://vcdn.bergfex.at/images/resized/7b/summary.jpg"
    assert result_page_1["summary_caption"] == "Summary Caption"

def test_parse_lelex_crozet_snow_data(lelex_crozet_html):
    data = parse_resort_page(lelex_crozet_html)

    assert data["resort_name"] == "Lélex - Crozet"
    assert data["new_snow"] == "15"
    assert data["snow_mountain"] == "15"
    
    try:
        from zoneinfo import ZoneInfo
    except ImportError:
        from backports.zoneinfo import ZoneInfo

    tz = ZoneInfo("Europe/Vienna")
    expected_dt = datetime(2025, 11, 5, 14, 40, tzinfo=tz)
    assert data["snow_valley"] == "5"
    assert data["lifts_open_count"] == 8
    assert data["lifts_total_count"] == 10
    assert data["status"] == "Open"
    assert data["last_update"] == expected_dt

def test_parse_overview_data_resort_name():
    """Test extracting resort name from overview table."""
    html = """
    <table class="snow">
        <tr><th>Resort</th><th>Valley</th><th>Mountain</th><th>New</th><th>Lifts</th><th>Update</th></tr>
        <tr>
            <td>
                <div class="h1">
                    <a href="/oesterreich/aichelberglifte/">Aichelberglifte Karlstift</a>
                </div>
            </td>
            <td data-value="10">10 cm</td>
            <td data-value="20">20 cm</td>
            <td data-value="0">0 cm</td>
            <td>
                <div class="icon-status icon-status1"></div>
                2/3
            </td>
            <td data-value="Heute, 09:00">Heute, 09:00</td>
        </tr>
    </table>
    """
    data = parse_overview_data(html)
    area_data = data.get("/oesterreich/aichelberglifte/")
    assert area_data is not None
    assert area_data["resort_name"] == "Aichelberglifte Karlstift"
    assert area_data["snow_valley"] == "10"
    assert area_data["snow_mountain"] == "20"
    assert area_data["status"] == "Open"
    assert area_data["lifts_open_count"] == 2


@pytest.fixture
def glacier3000_html():
    """Load Glacier 3000 - Les Diablerets fixture."""
    fixture_path = Path(__file__).parent / "fixtures" / "glacier3000.html"
    with open(fixture_path, "r", encoding="utf-8") as f:
        return f.read()


class TestGlacier3000Parsing:
    """Test parsing of Glacier 3000 - Les Diablerets page with all edge cases."""
    
    def test_slopes_open_km_with_german_decimal(self, glacier3000_html):
        """Verify slopes km are correctly parsed with German decimal notation (comma)."""
        data = parse_resort_page(glacier3000_html)
        
        # Should parse "10,8 km" as 10.8 (German decimal -> Python float)
        assert data["slopes_open_km"] == 10.8
        # Should parse "31,1 km" as 31.1
        assert data["slopes_total_km"] == 31.1
    
    def test_slope_condition_extracted(self, glacier3000_html):
        """Verify Pistenzustand (slope condition) is correctly extracted."""
        data = parse_resort_page(glacier3000_html)
        
        assert data["slope_condition"] == "gut"
    
    def test_last_snowfall_extracted(self, glacier3000_html):
        """Verify Letzter Schneefall Region is correctly extracted."""
        data = parse_resort_page(glacier3000_html)
        
        assert data["last_snowfall"] == "08.12.2025 - Tal: Heute"
    
    def test_last_update_with_desktop_only_class(self, glacier3000_html):
        """Verify last_update is parsed from div with 'desktop-only h2-sub' class."""
        try:
            from zoneinfo import ZoneInfo
        except ImportError:
            from backports.zoneinfo import ZoneInfo
        
        data = parse_resort_page(glacier3000_html)
        
        # Should parse "Heute, 10:46" correctly
        assert data["last_update"] is not None
        # Verify it's a datetime with correct time
        assert data["last_update"].hour == 10
        assert data["last_update"].minute == 46
    
    def test_snow_condition_extracted(self, glacier3000_html):
        """Verify Schneezustand (snow condition) is correctly extracted."""
        data = parse_resort_page(glacier3000_html)
        
        assert data["snow_condition"] == "Pulver"
    
    def test_resort_name_extracted(self, glacier3000_html):
        """Verify resort name is correctly extracted."""
        data = parse_resort_page(glacier3000_html)
        
        assert data["resort_name"] == "Glacier 3000 - Les Diablerets"
    
    def test_snow_depths_extracted(self, glacier3000_html):
        """Verify snow depths are correctly extracted."""
        data = parse_resort_page(glacier3000_html)
        
        assert data["snow_mountain"] == "381"
        # Note: snow_valley includes 'neu: X' text from nested div element
        # This is existing behavior - could be improved in future
        assert "170" in data["snow_valley"]
    
    def test_lifts_extracted(self, glacier3000_html):
        """Verify lift counts are correctly extracted."""
        data = parse_resort_page(glacier3000_html)
        
        assert data["lifts_open_count"] == 7
        assert data["lifts_total_count"] == 13
    
    def test_elevation_valley_extracted(self, glacier3000_html):
        """Verify valley elevation is extracted from '(Piste, 1.200m)'."""
        data = parse_resort_page(glacier3000_html)
        
        assert data["elevation_valley"] == 1200
    
    def test_new_snow_extracted(self, glacier3000_html):
        """Verify new snow is extracted from the heading-ne div."""
        data = parse_resort_page(glacier3000_html)
        
        assert data["new_snow"] == "5"
