requirements.txt
beautifulsoup4>=4.12.0
lxml>=5.1.0
requests>=2.31.0
pytest>=7.4.0
requests-mock>=1.11.0
google-cloud-storage>=2.14.0
google-cloud-bigquery>=3.17.0
python-dotenv>=1.0.0
functions-framework>=3.0.0
